+++ src/main/java/com/nexters/phochak/ignore/domain/IgnoredUsersRelation.java
@@ -1,5 +1,6 @@
-package com.nexters.phochak.domain;
+package com.nexters.phochak.ignore.domain;
 
+import com.nexters.phochak.domain.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
