+++ src/main/java/com/nexters/phochak/service/impl/UserServiceImpl.java
@@ -4,15 +4,15 @@ import com.nexters.phochak.auth.UserContext;
 import com.nexters.phochak.auth.application.OAuthService;
 import com.nexters.phochak.common.exception.PhochakException;
 import com.nexters.phochak.common.exception.ResCode;
-import com.nexters.phochak.domain.IgnoredUsers;
-import com.nexters.phochak.domain.IgnoredUsersRelation;
 import com.nexters.phochak.domain.User;
 import com.nexters.phochak.dto.OAuthUserInformation;
 import com.nexters.phochak.dto.response.IgnoredUserResponseDto;
 import com.nexters.phochak.dto.response.UserCheckResponseDto;
 import com.nexters.phochak.dto.response.UserInfoResponseDto;
+import com.nexters.phochak.ignore.domain.IgnoredUserRepository;
+import com.nexters.phochak.ignore.domain.IgnoredUsers;
+import com.nexters.phochak.ignore.domain.IgnoredUsersRelation;
 import com.nexters.phochak.notification.application.NotificationService;
-import com.nexters.phochak.repository.IgnoredUserRepository;
 import com.nexters.phochak.repository.UserRepository;
 import com.nexters.phochak.service.PostService;
 import com.nexters.phochak.service.UserService;
